Add timing of data loader and new defaults
Summary: data loading can introduce a large overhead, especially when
training small models. This commit adds code to time data loading and
also changes the default number of CPU workers per GPU to 8, which helps
training of small models when CPUs largely outnumber GPUs, as is common.

-benchmark.py: 
  -compute_time_eval()/train() moved here form net.py
  -compute_time_loader() new func to time data loader
  -compute_time_full() new func to time model+loader for iter+epochs
-config.py: DATA_LOADER.NUM_WORKERS set to 8
-net.py: moved timing function to benchmark.py
-trainer.py: uses benchmark.compute_time_full() 
-time_net.py: updated block doc and prints
