+++ pycls/core/config.py
@@ -288,8 +288,8 @@ _C.TEST.WEIGHTS = ""
 # ------------------------------------------------------------------------------------ #
 _C.DATA_LOADER = CfgNode()
 
-# Number of data loader workers per training process
-_C.DATA_LOADER.NUM_WORKERS = 4
+# Number of data loader workers per process
+_C.DATA_LOADER.NUM_WORKERS = 8
 
 # Load data to pinned host memory
 _C.DATA_LOADER.PIN_MEMORY = True
