+++ pycls/core/net.py
@@ -105,77 +105,6 @@ def complexity(model):
     return {"flops": cx["flops"], "params": cx["params"], "acts": cx["acts"]}
 
 
-@torch.no_grad()
-def compute_time_eval(model):
-    """Computes precise model forward test time using dummy data."""
-    # Use eval mode
-    model.eval()
-    # Generate a dummy mini-batch and copy data to GPU
-    im_size, batch_size = cfg.TRAIN.IM_SIZE, int(cfg.TEST.BATCH_SIZE / cfg.NUM_GPUS)
-    inputs = torch.zeros(batch_size, 3, im_size, im_size).cuda(non_blocking=False)
-    # Compute precise forward pass time
-    timer = Timer()
-    total_iter = cfg.PREC_TIME.NUM_ITER + cfg.PREC_TIME.WARMUP_ITER
-    for cur_iter in range(total_iter):
-        # Reset the timers after the warmup phase
-        if cur_iter == cfg.PREC_TIME.WARMUP_ITER:
-            timer.reset()
-        # Forward
-        timer.tic()
-        model(inputs)
-        torch.cuda.synchronize()
-        timer.toc()
-    return timer.average_time
-
-
-def compute_time_train(model, loss_fun):
-    """Computes precise model forward + backward time using dummy data."""
-    # Use train mode
-    model.train()
-    # Generate a dummy mini-batch and copy data to GPU
-    im_size, batch_size = cfg.TRAIN.IM_SIZE, int(cfg.TRAIN.BATCH_SIZE / cfg.NUM_GPUS)
-    inputs = torch.rand(batch_size, 3, im_size, im_size).cuda(non_blocking=False)
-    labels = torch.zeros(batch_size, dtype=torch.int64).cuda(non_blocking=False)
-    # Cache BatchNorm2D running stats
-    bns = [m for m in model.modules() if isinstance(m, torch.nn.BatchNorm2d)]
-    bn_stats = [[bn.running_mean.clone(), bn.running_var.clone()] for bn in bns]
-    # Compute precise forward backward pass time
-    fw_timer, bw_timer = Timer(), Timer()
-    total_iter = cfg.PREC_TIME.NUM_ITER + cfg.PREC_TIME.WARMUP_ITER
-    for cur_iter in range(total_iter):
-        # Reset the timers after the warmup phase
-        if cur_iter == cfg.PREC_TIME.WARMUP_ITER:
-            fw_timer.reset()
-            bw_timer.reset()
-        # Forward
-        fw_timer.tic()
-        preds = model(inputs)
-        loss = loss_fun(preds, labels)
-        torch.cuda.synchronize()
-        fw_timer.toc()
-        # Backward
-        bw_timer.tic()
-        loss.backward()
-        torch.cuda.synchronize()
-        bw_timer.toc()
-    # Restore BatchNorm2D running stats
-    for bn, (mean, var) in zip(bns, bn_stats):
-        bn.running_mean, bn.running_var = mean, var
-    return fw_timer.average_time, bw_timer.average_time
-
-
-def compute_time_full(model, loss_fun):
-    """Computes precise model time for both eval and train mode."""
-    test_fw_time = compute_time_eval(model)
-    train_fw_time, train_bw_time = compute_time_train(model, loss_fun)
-    return {
-        "test_fw_time": test_fw_time,
-        "train_fw_time": train_fw_time,
-        "train_bw_time": train_bw_time,
-        "train_fw_bw_time": train_fw_time + train_bw_time,
-    }
-
-
 def drop_connect(x, drop_ratio):
     """Drop connect (adapted from DARTS)."""
     keep_ratio = 1.0 - drop_ratio
