+++ tools/time_net.py
@@ -5,7 +5,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-"""Compute precise time for a model on 1 GPU."""
+"""Compute model and loader timings."""
 
 import pycls.core.config as config
 import pycls.core.distributed as dist
@@ -14,7 +14,7 @@ from pycls.core.config import cfg
 
 
 def main():
-    config.load_cfg_fom_args("Compute precise time for a model on 1 GPU.")
+    config.load_cfg_fom_args("Compute model and loader timings.")
     config.assert_and_infer_cfg()
     cfg.freeze()
     dist.multi_proc_run(num_proc=cfg.NUM_GPUS, fun=trainer.time_model)
